requirements.txt
streamlit
langchain-core
langgraph
langchain-google-genai
python-dotenv
yfinance
pandas
